AutoSearch-LLM-main/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
openai
httpx
beautifulsoup4
lxml
readability-lxml
